requirements.txt
# YouTube Analytics & Data APIs
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# AI/ML Integration - Gemini (Principal)
google-generativeai==0.3.2

# AI/ML Integration - Claude (Fallback)
anthropic==0.7.7

# AI/ML Integration - OpenAI (Fallback)
openai==1.3.5

# Google Cloud AI (Opcional)
google-cloud-aiplatform==1.38.1

# Data Processing & Analysis
pandas==2.1.3
numpy==1.25.2
python-dateutil==2.8.2

# Database & Storage
sqlalchemy==2.0.23

# Web Framework & API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0

# Data Visualization & Reporting
plotly==5.17.0
dash==2.14.2
matplotlib==3.8.2
seaborn==0.13.0

# Environment & Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Utilities
schedule==1.2.0
python-slugify==8.0.1
tqdm==4.66.1 